Collapse factory_database mod
This change collapses the factory_database module up into the root
factory module.
